fix deletion

+++ src/admin/admin.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import * as bcrypt from 'bcrypt';
 import { Product } from 'src/model/product.entity';
@@ -23,13 +23,27 @@ export class AdminService {
         return this.userRepository.find();
     }
 
+    async putUser(id: number, user: User) {
+        const updateUser = {...user};
+        if (user.user_password) {
+            updateUser.user_password = await bcrypt.hash(user.user_password, 10); 
+        } else if ('user_password' in updateUser) {
+            delete updateUser.user_password;
+        }
+        return await this.userRepository.update({id}, updateUser);
+    }
+
     async postUser(user: User) {
         const user_password = await bcrypt.hash(user.user_password, 10); 
         return await this.userRepository.insert({...user,user_password });
     }
 
     async deleteUser(id: number) {
-        return await this.userRepository.delete({id});
+        if (id === 1) { 
+            throw new ForbiddenException('Cannot delete admin user');
+        } else {
+            return await this.userRepository.delete({id});
+        }
     }
 
     postProduct(product: Product) {
